pyproject.toml
[project]
name = "boostapi"
version = "0.1.0"
description = "Production FastAPI starter toolkit with Redis, SQLAlchemy, JWT Auth, and CLI scaffolding"
readme = "README.md"
authors = [{name = "Dhinakaran", email = "dhinakaranthangaraj198@gmail.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
keywords = [
    "fastapi", "semantic-search", "pgvector", "redis", "vector-search",
    "embeddings", "nlp", "async", "postgresql", "sqlalchemy"
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.8.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "loguru>=0.7.0",
    "click>=8.1.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "aiosqlite>=0.20.0",
    "ruff>=0.6.0",
    "black>=24.0",
    "mypy>=1.11",
    "pre-commit>=3.7.0",
    "types-passlib>=1.7.7",
]
docker = ["docker>=7.0.0"]
ml = [
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
]

[project.urls]
Homepage = "https://github.com/dhinakaran/boostapi"
Documentation = "https://boostapi.readthedocs.io"
Repository = "https://github.com/dhinakaran/boostapi"
"Bug Tracker" = "https://github.com/dhinakaran/boostapi/issues"
Changelog = "https://github.com/dhinakaran/boostapi/blob/main/CHANGELOG.md"

[project.scripts]
boostapi = "boostapi.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/boostapi"]

[tool.hatch.version]
path = "src/boostapi/__init__.py"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "S",   # flake8-bandit
]
ignore = [
    "ANN101",  # Missing type annotation for `self`
    "ANN102",  # Missing type annotation for `cls`
    "S101",    # Use of assert
]

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "sqlalchemy"]

[tool.black]
target-version = ["py310", "py311", "py312"]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["migrations/"]

[[tool.mypy.overrides]]
module = "passlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jose.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/boostapi --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["src/boostapi"]
omit = ["*/migrations/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
